+++ RayTracing/Primitives/Plane.cs
@@ -7,12 +7,12 @@ namespace RayTracing.Primitives
     {
 		public Color Color { get; set; }
 	    public int Specular { get; set; }
-	    public float Reflect { get; set; }
+	    public double Reflect { get; set; }
 
-	    public float A { get; set; }
-		public float B { get; set; }
-		public float C { get; set; }
-		public float D { get; set; }
+	    public double A { get; set; }
+		public double B { get; set; }
+		public double C { get; set; }
+		public double D { get; set; }
 
 		//todo optimize dont create array, then replace abc to this
 	    public Vector Normal => new Vector(A, B, C);
